--- a/java/vector/src/main/java/org/apache/arrow/vector/BitVector.java
+++ b/java/vector/src/main/java/org/apache/arrow/vector/BitVector.java
@@ -120,16 +120,8 @@ public final class BitVector extends BaseFixedWidthVector {
   }
 
   @Override
-  protected void refreshValueCapacity() {
-    actualValueCapacity = Math.min(getValueBufferValueCapacity(), getValidityBufferValueCapacity());
-  }
-
-  private int getValueBufferValueCapacity() {
+  protected int getValueBufferValueCapacity() {
     return capAtMaxInt(valueBuffer.capacity() * 8);
-  }
-
-  private int getValidityBufferValueCapacity() {
-    return capAtMaxInt(validityBuffer.capacity() * 8);
   }
 
   /**