@josiahyan 2020-09-18
> I am thinking about another way: can we simply save the value/validity buffer capacities? so we can further improve the performance by avoiding the if branch in getValueBufferValueCapacity and the if branch in capAtMaxInt

Let me check again on this; I might be wrong about the performance differences and safety if we aggressively cache; it was something I gave less thought to because of our specific use case (`arrow.enable_unsafe_memory_access=false`), but it might be more beneficial to do in general.