@liyafan82 2020-09-21
@josiahyan Thank you for the additional details. 

I think one of your concern is that, the underlying buffers can be changed unintentionally, which lefts the vector in an inconsistent state (Please correct me if my understanding is incorrect, or if I have missed something). 

IMO, we encourage users to manipulate vectors through standard APIs (e.g. get/set/setSafe, etc. ). However, unintended manipulations are always possible, because the getXXXBuffer APIs are there. More fundamentally, the buffers are all based on off-heap memory, so if the user has the memory address, they can manipulate the buffer data at will. 

So I think what we should do is to design and implement the APIs to try making vectors in a consistent state, and if the vector falls into an inconsistent state, we should trap it with minimal cost.