@jacques-n 2020-09-21
@josiahyan , out of curiosity, were other patterns impactful to perf (further specialization say, per vector type) or were they basically the same improvement as this generic implementation?